move configuration utils
